+++ src/main/java/com/dto/CompanyDto.java
@@ -1,5 +1,7 @@
 package com.dto;
 
+import java.math.BigInteger;
+
 import com.model.UpdateType;
 
 import lombok.AllArgsConstructor;
@@ -11,7 +13,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class CompanyDto {
 	
-	private Long id;
+	private BigInteger id;
 	private String description;
 	private UpdateType updateType;
 }
