+++ src/main/java/com/dto/ProductsDto.java
@@ -1,5 +1,7 @@
 package com.dto;
 
+import java.math.BigInteger;
+
 import com.model.UpdateType;
 
 import lombok.AllArgsConstructor;
@@ -11,10 +13,10 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ProductsDto {
 
-	private Long id;
+	private BigInteger id;
 	private String codproduct;
 	private String description;
-	private Long id_company;
+	private BigInteger id_company;
 	private UpdateType updateType;
 	
 }
