+++ src/main/java/com/model/Products.java
@@ -1,7 +1,11 @@
 package com.model;
 
+import static javax.persistence.FetchType.LAZY;
 import static javax.persistence.GenerationType.SEQUENCE;
+
+import java.math.BigInteger;
 import java.time.Instant;
+
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
@@ -14,7 +18,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import static javax.persistence.FetchType.LAZY;
 
 @Data
 @AllArgsConstructor
@@ -26,7 +29,7 @@ public class Products {
 	@Id
 	@SequenceGenerator(name = "PRODUCTS_SEQ", sequenceName = "PRODUCTS_SEQ", initialValue = 1, allocationSize=10)
 	@GeneratedValue(strategy = SEQUENCE, generator = "PRODUCTS_SEQ")
-	private Long ProductId;
+	private BigInteger ProductId;
 	private String codproduct;
 	private String description;
 	private boolean enabled;
