+++ src/main/java/com/model/Purchases.java
@@ -3,6 +3,7 @@ package com.model;
 import static javax.persistence.FetchType.LAZY;
 import static javax.persistence.GenerationType.SEQUENCE;
 
+import java.math.BigInteger;
 import java.time.Instant;
 
 import javax.persistence.Entity;
@@ -11,7 +12,6 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.SequenceGenerator;
-import javax.persistence.Transient;
 import javax.persistence.Transient;
 
 import lombok.AllArgsConstructor;
@@ -29,7 +29,7 @@ public class Purchases {
 	@Id
 	@SequenceGenerator(name = "PURCHASES_SEQ", sequenceName = "PURCHASES_SEQ", initialValue = 1, allocationSize=10)
 	@GeneratedValue(strategy = SEQUENCE, generator = "PURCHASES_SEQ")
-	private Long PurchasesId;
+	private BigInteger PurchasesId;
 	private String description;
 	private double totalPrice;
 	private Instant createdate;
