+++ src/main/java/com/model/PurchasesDetails.java
@@ -3,6 +3,7 @@ package com.model;
 import static javax.persistence.FetchType.LAZY;
 import static javax.persistence.GenerationType.SEQUENCE;
 
+import java.math.BigInteger;
 import java.time.Instant;
 
 import javax.persistence.Entity;
@@ -11,8 +12,6 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.SequenceGenerator;
-import javax.persistence.Transient;
-
 import javax.persistence.Transient;
 
 import lombok.AllArgsConstructor;
@@ -30,7 +29,7 @@ public class PurchasesDetails {
 	@Id
 	@SequenceGenerator(name = "PURCHASESDETAILS_SEQ", sequenceName = "PURCHASESDETAILS_SEQ", initialValue = 1, allocationSize=10)
 	@GeneratedValue(strategy = SEQUENCE, generator = "PURCHASESDETAILS_SEQ")
-	private Long PurchasesDetId;
+	private BigInteger PurchasesDetId;
 	private String description;
 	private int quantity;
 	private Double unitaryCost;
