+++ src/main/java/com/model/VerificationToken.java
@@ -1,15 +1,20 @@
 package com.model;
 
+import static javax.persistence.FetchType.LAZY;
+import static javax.persistence.GenerationType.IDENTITY;
+
+import java.math.BigInteger;
+import java.time.Instant;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.OneToOne;
+import javax.persistence.Table;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-import java.time.Instant;
-
-import static javax.persistence.FetchType.LAZY;
-import static javax.persistence.GenerationType.IDENTITY;
-import static javax.persistence.GenerationType.SEQUENCE;
 
 @Data
 @AllArgsConstructor
@@ -20,7 +25,7 @@ public class VerificationToken {
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
-    private Long VerificationId;
+    private BigInteger VerificationId;
     private String token;
     @OneToOne(fetch = LAZY)
     private Users user;
