+++ src/main/java/com/repository/PurchasesRepository.java
@@ -1,5 +1,6 @@
 package com.repository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,8 +9,8 @@ import org.springframework.stereotype.Repository;
 import com.model.Purchases;
 
 @Repository
-public interface PurchasesRepository extends JpaRepository<Purchases, Long> {
+public interface PurchasesRepository extends JpaRepository<Purchases, BigInteger> {
 	
-	Optional<Purchases> findById(Long id);
+	Optional<Purchases> findById(BigInteger id);
 
 }
