+++ src/main/java/com/repository/RefreshTokenRepository.java
@@ -1,5 +1,6 @@
 package com.repository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.model.RefreshToken;
 
 @Repository
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, BigInteger> {
 	
 	Optional <RefreshToken> findByToken(String token);
