+++ src/main/java/com/repository/VerificationTokenRepository.java
@@ -1,13 +1,15 @@
 package com.repository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import com.model.VerificationToken;
 
 @Repository
-public interface VerificationTokenRepository extends JpaRepository<VerificationToken, Long> {
+public interface VerificationTokenRepository extends JpaRepository<VerificationToken, BigInteger> {
 	
 	Optional<VerificationToken> findByToken(String token);
